+++ src/reuse/_main.py
@@ -138,22 +138,7 @@ def parser() -> argparse.ArgumentParser:
         download.run,
         help=_("download a license and place it in the LICENSES/ directory"),
         description=fill_all(
-            _(
-                "Download a license and place it in the LICENSES/ directory.\n"
-                "\n"
-                "The LICENSES/ directory is automatically found in the"
-                " following order:\n"
-                "\n"
-                "- The LICENSES/ directory in the root of the VCS"
-                " repository.\n"
-                "\n"
-                "- The current directory if its name is LICENSES.\n"
-                "\n"
-                "- The LICENSES/ directory in the current directory.\n"
-                "\n"
-                "If the LICENSES/ directory cannot be found, one is simply"
-                " created."
-            )
+            _("Download a license and place it in the LICENSES/ directory.")
         ),
     )